requirements.txt
amqp==5.2.0
celery==5.3.4
Django==4.2.7
django-filter==23.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
pandas==2.1.2

